refactor(jsdoc): use Syntax from @jsdoc/parse
