+++ packages/jsdoc/package.json
@@ -17,6 +17,7 @@
     "@babel/parser": "^7.12.11",
     "@jsdoc/cli": "^0.2.5",
     "@jsdoc/core": "^0.4.0",
+    "@jsdoc/parse": "^0.1.0",
     "@jsdoc/util": "^0.2.4",
     "bluebird": "^3.7.2",
     "catharsis": "^0.9.0",
