+++ packages/jsdoc/test/specs/jsdoc/src/parser.js
@@ -188,7 +188,7 @@ describe('jsdoc/src/parser', () => {
             });
 
             it('should call AST node visitors', () => {
-                const Syntax = require('jsdoc/src/syntax').Syntax;
+                const { Syntax } = require('@jsdoc/parse');
 
                 let args;
                 const sourceCode = ['javascript:/** foo */var foo;'];
