--- packages/jsdoc/test/specs/jsdoc/src/syntax.js
@@ -1,7 +1,0 @@
-describe('jsdoc/src/syntax', () => {
-    const Syntax = require('jsdoc/src/syntax').Syntax;
-
-    it('should exist', () => {
-        expect(Syntax).toBeObject();
-    });
-});
